refactored config to split by domain to avoid config.js having too manyn lines
